# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ScanError"
            },
            "title": "Errors",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanQuality"
              },
              {
                "type": "null"
              }
            ],
            "description": "Evidence quality assessment. Determines whether AI fallback is recommended."
          },
          "scan": {
            "$ref": "#/components/schemas/ScanMetadata"
          },
          "schema_version": {
            "default": "2.0",
            "title": "Schema Version",
            "type": "string"
          },
          "screenshots": {
            "items": {
              "$ref": "#/components/schemas/ScreenshotRef"
            },
            "title": "Screenshots",
            "type": "array"
          },
          "sections": {
            "$ref": "#/components/schemas/SectionSet"
          },
          "target": {
            "$ref": "#/components/schemas/TargetInfo"
          }
        },
        "required": [
          "scan",
          "target",
          "sections"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntelligenceRequest": {
        "additionalProperties": false,
        "description": "Request body for the AI intelligence fallback.",
        "properties": {
          "additional_context": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional user-provided public context (engineering blog URLs, GitHub repos, etc.)",
            "title": "Additional Context"
          },
          "sections": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SectionKey"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sections to enhance with AI. None means all sections where AI is recommended.",
            "title": "Sections"
          }
        },
        "title": "IntelligenceRequest",
        "type": "object"
      },
      "IntelligenceResponse": {
        "additionalProperties": false,
        "description": "Response from the AI intelligence fallback.",
        "properties": {
          "findings_added": {
            "title": "Findings Added",
            "type": "integer"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "mode": {
            "default": "ai_intelligence_fallback",
            "description": "Analysis mode. Always 'ai_intelligence_fallback' for this endpoint.",
            "title": "Mode",
            "type": "string"
          },
          "quality_after": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanQuality"
              },
              {
                "type": "null"
              }
            ]
          },
          "quality_before": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanQuality"
              },
              {
                "type": "null"
              }
            ]
          },
          "research_available": {
            "title": "Research Available",
            "type": "boolean"
          },
          "research_performed": {
            "title": "Research Performed",
            "type": "boolean"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "string"
          },
          "sections_enhanced": {
            "items": {
              "$ref": "#/components/schemas/SectionKey"
            },
            "title": "Sections Enhanced",
            "type": "array"
          }
        },
        "required": [
          "scan_id",
          "sections_enhanced",
          "research_performed",
          "research_available",
          "findings_added"
        ],
        "title": "IntelligenceResponse",
        "type": "object"
      },
      "IntelligenceStatus": {
        "additionalProperties": false,
        "description": "Status of the AI intelligence run.",
        "properties": {
          "available": {
            "description": "Whether AI intelligence is configured and available.",
            "title": "Available",
            "type": "boolean"
          },
          "inference_available": {
            "description": "Whether an AI inference provider is configured.",
            "title": "Inference Available",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation when AI is not available.",
            "title": "Reason"
          },
          "research_available": {
            "description": "Whether a search provider is configured.",
            "title": "Research Available",
            "type": "boolean"
          }
        },
        "required": [
          "available",
          "research_available",
          "inference_available"
        ],
        "title": "IntelligenceStatus",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "RFC 9457 problem document.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the problem kind.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ScanAcceptedResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "normalized_url": {
            "title": "Normalized Url",
            "type": "string"
          },
          "requested_url": {
            "title": "Requested Url",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          }
        },
        "required": [
          "scan_id",
          "status",
          "requested_url",
          "normalized_url",
          "created_at",
          "links"
        ],
        "title": "ScanAcceptedResponse",
        "type": "object"
      },
...
      "ScanJobState": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "problem": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "$ref": "#/components/schemas/StageProgress"
          },
          "requested_url": {
            "title": "Requested Url",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageRun"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          }
        },
        "required": [
          "scan_id",
          "status",
          "requested_url"
        ],
        "title": "ScanJobState",
        "type": "object"
      },
...
      "ScanOptions": {
        "additionalProperties": false,
        "description": "Per-scan options. ``extra='forbid'`` so an option typo is a 422, not a silent no-op.",
        "properties": {
          "include_full_page_screenshot": {
            "default": false,
            "title": "Include Full Page Screenshot",
            "type": "boolean"
          },
          "include_screenshot": {
            "default": true,
            "title": "Include Screenshot",
            "type": "boolean"
          },
          "responsive_widths": {
            "items": {
              "type": "integer"
            },
            "maxItems": 6,
            "title": "Responsive Widths",
            "type": "array"
          },
          "sections": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SectionKey"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restrict analysis to these sections. None means all.",
            "title": "Sections"
          },
          "viewport": {
            "$ref": "#/components/schemas/Viewport"
          }
        },
        "title": "ScanOptions",
        "type": "object"
      },
...
      "ScanRequest": {
        "additionalProperties": false,
        "properties": {
          "options": {
            "$ref": "#/components/schemas/ScanOptions"
          },
          "url": {
            "description": "Public http(s) URL. Normalized and guarded server-side.",
            "maxLength": 2048,
            "minLength": 3,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ScanRequest",
        "type": "object"
      },
      "ScanStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "completed_with_errors",
          "failed",
          "cancelled"
        ],
        "title": "ScanStatus",
        "type": "string"
      },
...
      "SectionKey": {
        "description": "The four user-facing report sections. Order here is the canonical presentation order.\n\nV2 consolidated eight sections into four. Architecture, network, performance, SEO, and\naccessibility evidence still feeds the remaining sections internally but is no longer\nexposed as standalone user-facing reports.",
        "enum": [
          "design",
          "technology",
          "security",
          "traffic"
        ],
        "title": "SectionKey",
        "type": "string"
      },
...
      "StageProgress": {
        "additionalProperties": false,
        "description": "Progress derived from stages that actually completed.\n\nThere is no interpolation and no time-based estimate here on purpose: the requirement\nis honest progress, and the simplest way to guarantee it is to have no mechanism\ncapable of faking it.",
        "properties": {
          "completed_weight": {
            "default": 0,
            "title": "Completed Weight",
            "type": "integer"
          },
          "current_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageKey"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_stage_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage Label"
          },
          "stages_completed": {
            "default": 0,
            "title": "Stages Completed",
            "type": "integer"
          },
          "stages_total": {
            "default": 0,
            "title": "Stages Total",
            "type": "integer"
          },
          "total_weight": {
            "default": 0,
            "title": "Total Weight",
            "type": "integer"
          }
        },
        "title": "StageProgress",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Evidence-based website technical intelligence.\n\nDetection is deterministic and never performed by an AI model. Every asserted fact carries the\nevidence that supports it, and anything that cannot be established from observation is reported\nas not detected, not determinable, or unable to verify rather than guessed.\n\nAnalysis is passive: ReverseX observes what a normal visit reveals. It does not test\nauthentication, submit forms, fuzz inputs, or attempt to bypass access controls.",
    "title": "ReverseX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scans": {
      "post": {
        "operationId": "create_scan_api_v1_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a URL for analysis",
        "tags": [
          "scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}": {
      "delete": {
        "description": "Called by the client once the result is stored in IndexedDB.\n\nIdempotent: an unknown id also returns 204, because the desired end state - no server copy -\nholds either way.",
        "operationId": "delete_scan_api_v1_scans__scan_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release the server-side copy of a scan",
        "tags": [
          "scans"
        ]
      },
      "get": {
        "operationId": "get_scan_api_v1_scans__scan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanJobState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan status and stage progress",
        "tags": [
          "scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/events": {
      "get": {
        "operationId": "scan_events_api_v1_scans__scan_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Server-sent events for scan progress",
        "tags": [
          "scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/intelligence": {
      "post": {
        "description": "Trigger AI intelligence for sections with insufficient evidence.\n\nThis endpoint is user-initiated and never runs automatically. It:\n1. Reads the existing deterministic scan result\n2. Runs public web research (if a search provider is configured)\n3. Runs AI inference to fill gaps in the specified sections\n4. Updates the stored result with AI-inferred findings (clearly marked)\n\nAI findings NEVER overwrite or modify verified/deterministic findings.",
        "operationId": "run_intelligence_api_v1_scans__scan_id__intelligence_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntelligenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The scan has not finished yet."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "description": "No AI provider is configured."
          }
        },
        "summary": "Run AI intelligence fallback on an existing scan",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/intelligence/status": {
      "get": {
        "description": "Check whether AI intelligence can be triggered for this scan.\n\nReturns availability information without spending any API credits.",
        "operationId": "intelligence_status_api_v1_scans__scan_id__intelligence_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check if AI intelligence fallback is available",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/result": {
      "get": {
        "operationId": "get_result_api_v1_scans__scan_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The scan has not finished yet."
          },
          "410": {
            "description": "The result was released after the retention window."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Structured analysis result",
        "tags": [
          "scans"
        ]
      }
    },
...
  }
}